Check if input is long

+++ lab2/client/src/rsi/Pair.java
@@ -1,10 +1,10 @@
 package rsi;
 
 public class Pair {
-    int x;
-    int y;
+    long x;
+    long y;
 
-    public Pair(int x, int y) {
+    public Pair(long x, long y) {
         this.x = x;
         this.y = y;
     }
