+++ lab2/client/src/rsi/Task.java
@@ -4,5 +4,5 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface Task extends Remote {
-    ReturnType compute(int from, int to) throws RemoteException;
+    ReturnType compute(long from, long to) throws RemoteException;
 }
